+++ tests/testthat/test-objects.R
@@ -5,21 +5,26 @@ test_that("Patient..constructor", {
   on.exit({DBI::dbDisconnect(fake_db_conn, shutdown = TRUE)})
   
   dplyr::copy_to(fake_db_conn, patients, temporary = FALSE)
-  expect_error(Patient(fake_db_conn, NA))
-  expect_error(Patient(fake_db_conn, c()))
-  expect_error(Patient(fake_db_conn, c("a", "b")))
+  expect_error(Patient(fake_db_conn, NA),
+               "`id` must not be NA")
+  expect_error(Patient(fake_db_conn, c()),
+               "`id` must have length 1")
+  expect_error(Patient(fake_db_conn, c("a", "b")),
+               "`id` must have length 1")
   patient_object <- Patient(fake_db_conn, "99999999999")
   expect_s4_class(patient_object, "Patient")
   expect_s4_class(compute(patient_object), "Patient")
   expect_is(collect(patient_object), "tbl_df")
-  expect_error(Patient(fake_db_conn, 99999999999))
+  expect_error(Patient(fake_db_conn, 99999999999),
+               "`id` must be character")
   DBI::dbDisconnect(fake_db_conn, shutdown = TRUE)
   
   # works with integer/numeric
   patients <- dplyr::tibble(patient_id = 999)
   fake_db_conn <- DBI::dbConnect(duckdb::duckdb(), ":memory:", timezone_out = "UTC")
   dplyr::copy_to(fake_db_conn, patients, temporary = FALSE)
-  expect_error(Patient(fake_db_conn, "999"))
+  expect_error(Patient(fake_db_conn, "999"),
+               "`id` must be numeric")
   expect_s4_class(Patient(fake_db_conn, 999), "Patient")
   expect_s4_class(Patient(fake_db_conn, 999L), "Patient")
   DBI::dbDisconnect(fake_db_conn, shutdown = TRUE)
@@ -27,7 +32,8 @@ test_that("Patient..constructor", {
   patients <- dplyr::tibble(patient_id = 999L)
   fake_db_conn <- DBI::dbConnect(duckdb::duckdb(), ":memory:", timezone_out = "UTC")
   dplyr::copy_to(fake_db_conn, patients, temporary = FALSE)
-  expect_error(Patient(fake_db_conn, "999"))
+  expect_error(Patient(fake_db_conn, "999"),
+               "`id` must be integer")
   expect_s4_class(Patient(fake_db_conn, 999), "Patient")
   expect_s4_class(Patient(fake_db_conn, 999L), "Patient")
 })
@@ -173,12 +179,17 @@ test_that("MedicationRequest..constructor", {
                  dplyr::tibble(prescription_id = "999999"),
                  "drug_prescriptions", 
                  temporary = FALSE)
-  expect_error(MedicationRequest(fake_db_conn, NA))
-  expect_error(MedicationRequest(fake_db_conn, c()))
-  expect_warning(MedicationRequest(fake_db_conn, c("a", "b")))
+  on.exit({DBI::dbDisconnect(fake_db_conn, shutdown = TRUE)})
+  expect_error(MedicationRequest(fake_db_conn, NA),
+               "`id` must not be NA")
+  expect_error(MedicationRequest(fake_db_conn, c()),
+               "`id` must have length 1")
+  expect_warning(MedicationRequest(fake_db_conn, c("a", "b")),
+                 "`id` must have length 1")
   object <- MedicationRequest(fake_db_conn, "999999")
   expect_s4_class(object, "MedicationRequest")
-  expect_error(MedicationRequest(fake_db_conn, 999999))
+  expect_error(MedicationRequest(fake_db_conn, 999999),
+               "`id` must be character")
   DBI::dbDisconnect(fake_db_conn, shutdown = TRUE)
   
   # works with integer/numeric
@@ -187,7 +198,8 @@ test_that("MedicationRequest..constructor", {
                  dplyr::tibble(prescription_id = 999L),
                  "drug_prescriptions", 
                  temporary = FALSE)
-  expect_error(MedicationRequest(fake_db_conn, "999"))
+  expect_error(MedicationRequest(fake_db_conn, "999"),
+               "`id` must be integer")
   expect_s4_class(MedicationRequest(fake_db_conn, 999), "MedicationRequest")
   expect_s4_class(MedicationRequest(fake_db_conn, 999L), "MedicationRequest")
   DBI::dbDisconnect(fake_db_conn, shutdown = TRUE)
@@ -197,10 +209,10 @@ test_that("MedicationRequest..constructor", {
                  dplyr::tibble(prescription_id = 999),
                  "drug_prescriptions", 
                  temporary = FALSE)
-  expect_error(MedicationRequest(fake_db_conn, "999"))
+  expect_error(MedicationRequest(fake_db_conn, "999"),
+               "`id` must be numeric")
   expect_s4_class(MedicationRequest(fake_db_conn, 999), "MedicationRequest")
   expect_s4_class(MedicationRequest(fake_db_conn, 999L), "MedicationRequest")
-  DBI::dbDisconnect(fake_db_conn, shutdown = TRUE)
 })
 
 
@@ -382,33 +394,42 @@ test_that("MedicationRequest..interface_methods Postgres", {
 
 test_that("TherapyEpisode..constructor", {
   fake_db_conn <- DBI::dbConnect(duckdb::duckdb(), ":memory:", timezone_out = "UTC")
+  on.exit({DBI::dbDisconnect(fake_db_conn, shutdown = TRUE)})
   dplyr::copy_to(fake_db_conn, 
-                 dplyr::tibble(therapy_id = "999999"),
+                 dplyr::tibble(
+                   patient_id = "9",
+                   therapy_id = "999999"
+                 ),
                  "drug_therapy_episodes", 
                  temporary = FALSE)
-  expect_error(TherapyEpisode(fake_db_conn, NA))
-  expect_error(TherapyEpisode(fake_db_conn, c()))
-  expect_error(TherapyEpisode(fake_db_conn, 999999))
-  expect_error(TherapyEpisode(fake_db_conn, 999999L))
+  expect_error(TherapyEpisode(fake_db_conn, NA),
+               "`id` must contain at least one identifier")
+  expect_error(TherapyEpisode(fake_db_conn, c()),
+               "`id` must contain at least one identifier")
+  expect_error(TherapyEpisode(fake_db_conn, 999999),
+               "`id` must be character")
+  expect_error(TherapyEpisode(fake_db_conn, 999999L),
+               "`id` must be character")
   DBI::dbDisconnect(fake_db_conn, shutdown = TRUE)
   
   # works with integer/numeric
   fake_db_conn <- DBI::dbConnect(duckdb::duckdb(), ":memory:", timezone_out = "UTC")
   dplyr::copy_to(fake_db_conn, 
-                 dplyr::tibble(therapy_id = 999999L),
+                 dplyr::tibble(patient_id = 9L,
+                               therapy_id = 999999L),
                  "drug_therapy_episodes", 
                  temporary = FALSE)
-  expect_error(TherapyEpisode(fake_db_conn, "999999"))
+  expect_error(TherapyEpisode(fake_db_conn, "999999"),
+               "`id` must be integer")
   DBI::dbDisconnect(fake_db_conn, shutdown = TRUE)
   
   fake_db_conn <- DBI::dbConnect(duckdb::duckdb(), ":memory:", timezone_out = "UTC")
   dplyr::copy_to(fake_db_conn, 
-                 dplyr::tibble(therapy_id = 999999),
+                 dplyr::tibble(patient_id = 9,
+                               therapy_id = 999999),
                  "drug_therapy_episodes", 
                  temporary = FALSE)
   expect_error(TherapyEpisode(fake_db_conn, "999999"))
-  
-  DBI::dbDisconnect(fake_db_conn, shutdown = TRUE)
 })
 
 
@@ -578,33 +599,230 @@ test_that("TherapyEpisode..interface_methods Postgres", {
   DBI::dbDisconnect(conPostgreSQL)
 })
 
-test_that("TherapyEpisode..constructor", {
+test_that("Encounter..constructor", {
   fake_db_conn <- DBI::dbConnect(duckdb::duckdb(), ":memory:", timezone_out = "UTC")
-  dplyr::copy_to(fake_db_conn, 
-                 dplyr::tibble(therapy_id = "999999"),
-                 "drug_therapy_episodes", 
-                 temporary = FALSE)
-  expect_error(TherapyEpisode(fake_db_conn, NA))
-  expect_error(TherapyEpisode(fake_db_conn, c()))
-  expect_error(TherapyEpisode(fake_db_conn, 999999))
-  expect_error(TherapyEpisode(fake_db_conn, 999999L))
+  on.exit({DBI::dbDisconnect(fake_db_conn, shutdown = TRUE)})
+  dplyr::copy_to(
+    fake_db_conn, 
+    dplyr::tibble(
+      patient_id = "9",
+      encounter_id = "999999"
+    ),
+    "inpatient_episodes", 
+    temporary = FALSE
+  )
+  expect_error(Encounter(fake_db_conn, NA),
+               "`id` must contain at least one identifier")
+  expect_error(Encounter(fake_db_conn, c()),
+               "`id` must contain at least one identifier")
+  expect_error(Encounter(fake_db_conn, 999999),
+               "`id` must be character")
+  expect_error(Encounter(fake_db_conn, 999999L),
+               "`id` must be character")
+  expect_s4_class(Encounter(fake_db_conn, "999999"), "Encounter")
   DBI::dbDisconnect(fake_db_conn, shutdown = TRUE)
   
   # works with integer/numeric
   fake_db_conn <- DBI::dbConnect(duckdb::duckdb(), ":memory:", timezone_out = "UTC")
-  dplyr::copy_to(fake_db_conn, 
-                 dplyr::tibble(therapy_id = 999999L),
-                 "drug_therapy_episodes", 
-                 temporary = FALSE)
-  expect_error(TherapyEpisode(fake_db_conn, "999999"))
+  dplyr::copy_to(
+    fake_db_conn, 
+    dplyr::tibble(
+      patient_id = 9,
+      encounter_id = 999999
+    ),
+    "inpatient_episodes", 
+    temporary = FALSE
+  )
+  expect_error(Encounter(fake_db_conn, "999999"), 
+               "`id` must be numeric")
+  expect_s4_class(Encounter(fake_db_conn, 999999), "Encounter")
+  expect_s4_class(Encounter(fake_db_conn, 999999L), "Encounter")
   DBI::dbDisconnect(fake_db_conn, shutdown = TRUE)
   
   fake_db_conn <- DBI::dbConnect(duckdb::duckdb(), ":memory:", timezone_out = "UTC")
-  dplyr::copy_to(fake_db_conn, 
-                 dplyr::tibble(therapy_id = 999999),
-                 "drug_therapy_episodes", 
-                 temporary = FALSE)
-  expect_error(TherapyEpisode(fake_db_conn, "999999"))
-  DBI::dbDisconnect(fake_db_conn, shutdown = TRUE)
+  dplyr::copy_to(
+    fake_db_conn, 
+    dplyr::tibble(
+      patient_id = 9L,
+      encounter_id = 999999L
+    ),
+    "inpatient_episodes", 
+    temporary = FALSE
+  )
+  expect_error(Encounter(fake_db_conn, "999999"),
+               "`id` must be integer")
+  expect_s4_class(Encounter(fake_db_conn, 999999), "Encounter")
+  expect_s4_class(Encounter(fake_db_conn, 999999L), "Encounter")
 })
 
+
+test_that("Encounter..interface_methods DuckDB", {
+  conDuckDB <- DBI::dbConnect(duckdb::duckdb(), ":memory:", timezone_out = "UTC")
+  on.exit({DBI::dbDisconnect(conDuckDB, shutdown = TRUE)})
+  
+  fake_encounters <- dplyr::tibble(
+    patient_id = 6145252493, 
+    encounter_id = 5458286195:5458286199, 
+    admission_method = "1", 
+    admission_date = structure(1443082402, tzone = "UTC", class = c("POSIXct", "POSIXt")), 
+    discharge_date = structure(1443118601, tzone = "UTC", class = c("POSIXct", "POSIXt")), 
+    episode_number = 1L, 
+    last_episode_in_encounter = 1, 
+    episode_start = structure(1443082402, tzone = "UTC", class = c("POSIXct", "POSIXt")), 
+    episode_end = structure(1443118601, tzone = "UTC", class = c("POSIXct", "POSIXt")), 
+    consultant_code = "C1000003", 
+    main_specialty_code = "100"
+  )
+  load_inpatient_episodes(
+    conn = conDuckDB,
+    patients_data = dplyr::tibble(patient_id = 6145252493),
+    episodes_data = fake_encounters,
+    overwrite = TRUE
+  )
+  encounter_object <- Encounter(conDuckDB, 5458286195)
+  
+  # CLASS
+  expect_equal(
+    class(encounter_object),
+    structure("Encounter", package = "Ramses")
+  )
+  
+  # SHOW
+  expect_equal(
+    capture.output(encounter_object)[1:2],
+    c("Encounter 5458286195 ", "Patient:   6145252493 ")
+  )
+  expect_equal(
+    capture.output(Encounter(conDuckDB, 5458286195:5458286196))[1:3],
+    c("Encounters 5458286195, 5458286196 ", "[total of 2 encounters]", "Patient(s):   6145252493 ")
+  )
+  expect_equal(
+    capture.output(Encounter(conDuckDB, 5458286195:5458286199))[1:3],
+    c("Encounters 5458286195, 5458286196, 5458286197 ...", 
+      "[total of 5 encounters]", 
+      "Patient(s):   6145252493 ")
+  )
+  
+  # COMPUTE 
+  expect_equal(
+    encounter_object@record$lazy_query$x$x,
+    structure("inpatient_episodes", class = c("ident", "character"))
+  )
+  encounter_object_computed <- compute(encounter_object)
+  expect_true(
+    grepl("^dbplyr_", 
+          as.character(
+            encounter_object_computed@record$lazy_query$x
+          ))
+  )
+  
+  # COLLECT
+  expect_equal(
+    collect(encounter_object),
+    dplyr::tibble(
+      patient_id = 6145252493, 
+      encounter_id = 5458286195, 
+      admission_method = "1", 
+      admission_date = structure(1443082402, class = c("POSIXct", "POSIXt"), tzone = "UTC"), 
+      discharge_date = structure(1443118601, class = c("POSIXct", "POSIXt"), tzone = "UTC"), 
+      episode_number = 1L, 
+      last_episode_in_encounter = 1, 
+      episode_start = structure(1443082402, class = c("POSIXct", "POSIXt"), tzone = "UTC"), 
+      episode_end = structure(1443118601, class = c("POSIXct", "POSIXt"), tzone = "UTC"), 
+      consultant_code = "C1000003", 
+      main_specialty_code = "100", 
+      ramses_bed_days = 0.418969907407407
+    )
+  )
+})
+
+
+
+test_that("Encounter..interface_methods Postgres", {
+  
+  if (!identical(Sys.getenv("CI"), "true")) {
+    skip("Test only on Travis")
+  }
+  
+  conPostgreSQL <- DBI::dbConnect(RPostgres::Postgres(),
+                                  user = "user", 
+                                  password = "password",
+                                  host = "localhost", 
+                                  dbname="RamsesDB",
+                                  timezone = "UTC")
+  
+  fake_encounters <- dplyr::tibble(
+    patient_id = 6145252493, 
+    encounter_id = 5458286195:5458286199, 
+    admission_method = "1", 
+    admission_date = structure(1443082402, tzone = "UTC", class = c("POSIXct", "POSIXt")), 
+    discharge_date = structure(1443118601, tzone = "UTC", class = c("POSIXct", "POSIXt")), 
+    episode_number = 1L, 
+    last_episode_in_encounter = 1, 
+    episode_start = structure(1443082402, tzone = "UTC", class = c("POSIXct", "POSIXt")), 
+    episode_end = structure(1443118601, tzone = "UTC", class = c("POSIXct", "POSIXt")), 
+    consultant_code = "C1000003", 
+    main_specialty_code = "100"
+  )
+  load_inpatient_episodes(
+    conn = conPostgreSQL,
+    patients_data = dplyr::tibble(patient_id = 6145252493),
+    episodes_data = fake_encounters,
+    overwrite = TRUE
+  )
+  encounter_object <- Encounter(conPostgreSQL, 5458286195)
+  
+  # CLASS
+  expect_equal(
+    class(encounter_object),
+    structure("Encounter", package = "Ramses")
+  )
+  
+  # SHOW
+  expect_equal(
+    capture.output(encounter_object)[1:2],
+    c("Encounter 5458286195 ", "Patient:   6145252493 ")
+  )
+  expect_equal(
+    capture.output(Encounter(conPostgreSQL, 5458286195:5458286196))[1:3],
+    c("Encounters 5458286195, 5458286196 ", "[total of 2 encounters]", "Patient(s):   6145252493 ")
+  )
+  expect_equal(
+    capture.output(Encounter(conPostgreSQL, 5458286195:5458286199))[1:3],
+    c("Encounters 5458286195, 5458286196, 5458286197 ...", 
+      "[total of 5 encounters]", 
+      "Patient(s):   6145252493 ")
+  )
+  
+  # COMPUTE 
+  expect_equal(
+    encounter_object@record$lazy_query$x$x,
+    structure("inpatient_episodes", class = c("ident", "character"))
+  )
+  encounter_object_computed <- compute(encounter_object)
+  expect_true(
+    grepl("^dbplyr_", 
+          as.character(
+            encounter_object_computed@record$lazy_query$x
+          ))
+  )
+  
+  # COLLECT
+  expect_equal(
+    collect(encounter_object),
+    dplyr::tibble(
+      patient_id = 6145252493, 
+      encounter_id = 5458286195, 
+      admission_method = "1", 
+      admission_date = structure(1443082402, class = c("POSIXct", "POSIXt"), tzone = "UTC"), 
+      discharge_date = structure(1443118601, class = c("POSIXct", "POSIXt"), tzone = "UTC"), 
+      episode_number = 1L, 
+      last_episode_in_encounter = 1, 
+      episode_start = structure(1443082402, class = c("POSIXct", "POSIXt"), tzone = "UTC"), 
+      episode_end = structure(1443118601, class = c("POSIXct", "POSIXt"), tzone = "UTC"), 
+      consultant_code = "C1000003", 
+      main_specialty_code = "100", 
+      ramses_bed_days = 0.418969907407407
+    )
+  )
+})